chore: rm setup.sh
Since the programs that necessitated the automatic generation of a
script for running fpm were moved to the demo/ subdirectory,
retaining the setup.sh script obscures the fact that we can now
support the building, testing, and running of progtrams by entering
uncomplicated fpm commands such as `fpm test`.  This commit
therefore removes the script.