Move dynamic stack allocations outside `do concurrent`
Dyanmic stack alloctaions are those allocs where memory is allocated on
the stack (i.e. inside a function or proceure) with a size not known at
compile time.

Even though this is legal both at compile and run-times for the CPU on
the GPU this is not supported.

Therefore, this PR removes an instance of such allocations (this is the
only instance I came across so far while compiling for the GPU).