Remove ifx-only block which has invalid code
The use of ALL is invalid because it requires an array argument and
these argumens are scalar. It looks like this ifx code is supposed to do
the same thing as the non-ifx code but without DO CONCURRENT. I'm not
sure why. But the ifx code is wrong so it either needs to be fixed or
removed. This commit goes with the remove option.