fix(histogram_t):always define frequency component
This commit fixes an issue that was most likely introduced when
the bin-filling algorithm was bifurcated into one version below
a performance threshold of 80 bins and another version at or above
the threshold.  The histogram_t bin_frequency_ component was being
defined only above the threshold.  This commit defines the
component outside (after) the performance-threshold conditional
branching so the definition happens regardless of which branch
executes.