refac(example):simple learn-saturated-mixing-ratio
This commit adds a trainable_network_t constructor that accepts the
components derived types components directly rather than exposing
the use of its neural_network_t parent as an intermediary.  This
eliminates the need for the neural_network_t type in the
learn-saturated-mixing-ratio example and thereby makes that
example's use of Fiats entities a subset of the entities used by
the similar but much more complicated demonstration application
train-cloud-microphysics.  This change makes the example a useful
proxy for the demonstration application when introducing new users
to the demonstration application in the ISS conference paper.