build(include/language-support.F90): rm macro
This commit removes the HAVE_PROCEDURE_ACTUAL_FOR_POINTER_DUMMY
macro, which was used to work around a Fortran 2008 missing from
gfortran versions earlier than 14.3.  Because of reported gfortran
issues with parameterized derived types, it is unlikely that
future releases of Fiats will be able to support older versions of
gfortran so there's no reason to keep the workaround in place.